+++ history.md
@@ -14,6 +14,7 @@
 ### 4.6.2-ing
 - 升级data依赖
 - 适配MenuValue.confirm/hidden
+- 适配TableMenu.disabled
 - 适配ListEdit
 
 ### 4.6.1